requirements.txt
#requirements
pep8
django==1.5
django-webtest==1.7
WebTest==2.0
South==0.8
Pillow==2.1
django-ajax-upload-widget==0.5
